+++ src/bigxml/__init__.py
@@ -1,8 +1,10 @@
+from bigxml.exceptions import BigXmlException
 from bigxml.handler_marker import HandlerTypeHelper, xml_handle_element, xml_handle_text
 from bigxml.nodes import XMLElement, XMLText
 from bigxml.parser import Parser
 
 __all__ = (
+    "BigXmlException",
     "HandlerTypeHelper",
     "Parser",
     "xml_handle_element",