+++ tests/integration/test_encodings.py
@@ -1,9 +1,8 @@
 from typing import Iterator
-import xml.etree.ElementTree as ET
 
 import pytest
 
-from bigxml import Parser, XMLElement, xml_handle_element
+from bigxml import BigXmlException, Parser, XMLElement, xml_handle_element
 
 TXT_STR = "aàâæcçeéèêëiïîoôuùûü"
 XML_STR = f"<élément>{TXT_STR}</élément>"
@@ -67,5 +66,7 @@ def test_encoding(xml: bytes) -> None:
 def test_wrong_explicit_encoding() -> None:
     xml = ("<?xml version='1.0' encoding='ISO-8859-1'?>" + XML_STR).encode("utf_8")
     parser = Parser(xml)
-    with pytest.raises(ET.ParseError):
+    with pytest.raises(BigXmlException) as exc_info:
         parser.return_from()
+    assert str(exc_info.value) == "Not well-formed (invalid token): line 1, column 45"
+    assert not exc_info.value.security